requirements.txt
ffmpeg-python
imageio-ffmpeg
Pillow
controlnet_hinter